Automatically derive type instances
